web-app/backend/requirements.txt
# ── Core Framework ───────────────────────────────────────────────────────────
flask
flask-cors

# ── AI / OpenAI ──────────────────────────────────────────────────────────────
openai
anthropic
google-generativeai

# ── Database: PostgreSQL + pgvector (replaces sqlite3 + chromadb) ────────────
psycopg2-binary
pgvector

# ── Auth ──────────────────────────────────────────────────────────────────────
PyJWT
bcrypt

# ── Config / Utils ────────────────────────────────────────────────────────────
python-dotenv

# ── Document Parsing ──────────────────────────────────────────────────────────
PyPDF2
python-docx
openpyxl

# ── MCP Protocol (optional, requires Python 3.10+) ───────────────────────────
# mcp

# ── HTTP ──────────────────────────────────────────────────────────────────────
requests
